refactor(FS-2): refactor domain models and ports to use uint64 instead of int

+++ internal/core/domain/product.go
@@ -6,11 +6,11 @@ import (
 )
 
 type Product struct {
-  ID int 
+  ID uint64 
   Name string 
   Description string 
   Price float64 
-  CategoryID int
+  CategoryID uint64
   CreatedAt time.Time 
   UpdatedAt time.Time 
   DeletedAt time.Time
